% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_plan_stats.R
\name{add_plan_stats}
\alias{add_plan_stats}
\title{Add summary statistics to plans}
\usage{
add_plan_stats(plans, ref_plan, map = NULL, name = NULL, ...)
}
\arguments{
\item{plans}{a \code{redist_plans} object}

\item{ref_plan}{an integer vector of district assignments}

\item{map}{a \code{redist_map} object. Required if \code{plans} contains summary statistics.}

\item{name}{name for the reference plan. Defaults to \code{"ref"}.}

\item{...}{additional arguments (currently ignored)}
}
\value{
a \code{redist_plans} object with the reference plan added
}
\description{
Add summary statistics to plans
}
\examples{
dc_map <- redist::redist_map(dc, existing_plan = ward)
plans <- redist::redist_smc(dc_map, nsims = 10)
plans <- add_plan_stats(plans, dc_map$ward, map = dc_map, name = 'example')
}
