% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_editor.R
\name{adj_editor}
\alias{adj_editor}
\title{Interactive Adjacency Graph Editing}
\usage{
adj_editor(shp, adj = geomander::adjacency(shp), opts = redistio_options())
}
\arguments{
\item{shp}{an \code{sf} tibble that you want to draw with}

\item{adj}{a zero-indexed adjacency graph}

\item{opts}{list of options. Default is \code{redistio_options()}}
}
\value{
Shiny app
}
\description{
Interactive Adjacency Graph Editing
}
\examples{
if (interactive()) {
  adj_editor(dc)
}
}
