% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{guess_admins}
\alias{guess_admins}
\title{Guess which columns contain administrative units}
\usage{
guess_admins(shp)
}
\arguments{
\item{shp}{an \code{sf} tibble that you want to draw with}
}
\value{
a named \code{list} of types
}
\description{
Guess which columns contain administrative units
}
\examples{
guess_admins(dc)
}
