% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess.R
\name{guesstimate_demographics}
\alias{guesstimate_demographics}
\title{Guess and estimate which columns contain demographic data}
\usage{
guesstimate_demographics(shp)
}
\arguments{
\item{shp}{an \code{sf} tibble that you want to draw with}
}
\value{
a named \code{list} of columns
}
\description{
Guess and estimate which columns contain demographic data
}
\examples{
guesstimate_demographics(dc)
}
