% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistio_options.R
\name{redistio_options}
\alias{redistio_options}
\title{Set options for \code{redistio}}
\usage{
redistio_options(theme = "flatly", select_color = "purple", ...)
}
\arguments{
\item{theme}{a name of a bootswatch preset theme or other \code{bslib::bs_theme()} object}

\item{select_color}{a color to use for highlighting selected districts}

\item{...}{additional arguments (currently ignored)}
}
\value{
a \code{list}
}
\description{
Set options for \code{redistio}
}
\examples{
redistio_options()
}
