% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_editor.R
\name{adj_editor}
\alias{adj_editor}
\title{Interactive Adjacency Graph Editing}
\usage{
adj_editor(
  shp,
  adj = geomander::adjacency(shp),
  init_plan,
  palette = NULL,
  layers = NULL,
  hover_fn = hover_precinct,
  opts = redistio_options()
)
}
\arguments{
\item{shp}{an \code{sf} tibble that you want to draw with}

\item{adj}{a zero-indexed adjacency graph}

\item{init_plan}{plan to initialize coloring}

\item{palette}{Color palette to fill shapes with. Default is \verb{Polychrome 36} or,
if installed, \code{crayons::crayons$no_48}.}

\item{layers}{Named list of \code{sf} objects to overlay. Also takes column names in \code{shp} to group by.}

\item{hover_fn}{Function to generate tables for mouse hovering. Default is \code{hover_precinct()}.}

\item{opts}{list of options. Default is \code{redistio_options()}}
}
\value{
Shiny app
}
\description{
Interactive Adjacency Graph Editing
}
\examples{
if (interactive()) {
  adj_editor(dc, init_plan = dc$ward)
  adj_editor(dc, init_plan = dc$ward, layers = list(neighborhoods = 'adv_nbr'))
}
}
