% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{draw}
\alias{draw}
\title{Interactive Plan Drawing}
\usage{
draw(
  shp,
  init_plan,
  ndists,
  palette,
  pop_tol = 0.05,
  save_path = tempfile(fileext = ".baf")
)
}
\arguments{
\item{shp}{\code{sf} tibble that you want to draw with}

\item{init_plan}{Plan to initialize with.}

\item{ndists}{Number of districts to draw if \code{init_plan} is not supplied.}

\item{palette}{Color palette to fill shapes with. Default is Polychrome 36.}

\item{pop_tol}{the population tolerance.}

\item{save_path}{Output path to save progress to.}
}
\value{
Shiny app
}
\description{
Interactive Plan Drawing
}
\examples{
if (interactive()) {
  draw(dc, dc$ward)
}

}
