% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hover_precinct.R
\name{hover_precinct}
\alias{hover_precinct}
\title{Create a tibble of precinct stats}
\usage{
hover_precinct(shp, id, ...)
}
\arguments{
\item{shp}{\code{sf} tibble}

\item{id}{integer row number}

\item{...}{named tidyselections}
}
\value{
A \link[tibble:tibble]{tibble::tibble}
}
\description{
Create a tibble of precinct stats
}
\examples{
hover_precinct(dc, 1, pop = dplyr::starts_with('pop'), vap = dplyr::starts_with('vap'))
}
