% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistio_options.R
\name{redistio_options}
\alias{redistio_options}
\title{Set options for \code{redistio}}
\usage{
redistio_options(
  theme = "flatly",
  panels = c("elections", "demographics", "integrity", "algorithms", "plans",
    "comparisons"),
  select_color = "purple",
  border_color = "#000000",
  palette_pop = "Purples",
  palette_pct = "PuOr",
  palette_party = ggredist::ggredist$partisan,
  map_tiles = mapgl::carto_style("voyager"),
  leaflet_height = "91vh",
  crs = 4326,
  na_color = "#0000",
  layer_weight = 1.5,
  layer_color = "#000000",
  locked_districts = NULL,
  use_algorithms = TRUE,
  alg_max_districts = 3,
  alg_max_sims = 100,
  use_plans = TRUE,
  use_comparisons = TRUE,
  use_planscore = TRUE,
  save_assignment_path = "redistio.csv",
  save_shape_path = "redistio.geojson",
  debounce = 150,
  projection = "mercator",
  plot_theme = ggplot2::theme_bw(),
  plot_geom = "boxplot",
  plot_ref_geom = NULL,
  ...
)
}
\arguments{
\item{theme}{a name of a bootswatch preset theme or other \code{bslib::bs_theme()} object}

\item{panels}{which panels to display in the app, \code{'draw'} is always shown.}

\item{select_color}{a color to use for highlighting selected districts}

\item{border_color}{a color to use for precinct borders in the map}

\item{palette_pop}{a color palette to use for whole people. Defaults to
\code{RColorBrewer::brewer.pal(n = 3, name = 'Purples')}.}

\item{palette_pct}{a color palette to use for percentages of people. Defaults
to \code{RColorBrewer::brewer.pal(n = 3, name = 'PuOr')}.}

\item{palette_party}{a color palette to use for parties}

\item{map_tiles}{a tileset to use for the map background, from \code{leaflet::providers}}

\item{leaflet_height}{height to pass to \code{leaflet::leafletOutput()}}

\item{crs}{a coordinate reference system to use in \code{leaflet::leaflet()}}

\item{na_color}{a color to use for unassigned precincts}

\item{layer_weight}{a stroke width to use for layers in \code{leaflet::leaflet()}}

\item{layer_color}{colors to use for layers in \code{mapgl::maplibre()}}

\item{locked_districts}{districts to lock on app start to stop edits}

\item{use_algorithms}{whether to use redistricting simulation algorithms}

\item{alg_max_districts}{maximum number of districts to use in algorithms}

\item{alg_max_sims}{maximum number of simulations to use in algorithms}

\item{use_plans}{whether to show the plans browser tab when \code{plans} is provided}

\item{use_comparisons}{whether to show the comparisons tab when \code{plans} and \code{plans_fn} are provided}

\item{use_planscore}{whether to use PlanScore to evaluate plans}

\item{save_assignment_path}{Output path to save assignment file to.}

\item{save_shape_path}{Output path to save shapefile to.}

\item{debounce}{Number of milliseconds to debounce map hover events. Defaults
to 150.}

\item{projection}{Maplibre projection to use. Default is \code{'mercator'}.}

\item{plot_theme}{a ggplot2 theme to apply to comparison plots. Default is
\code{ggplot2::theme_bw()}.}

\item{plot_geom}{a geom to use for district-level comparison plots, passed
to the \code{geom} argument of \code{\link[redist:redist.plot.distr_qtys]{redist::redist.plot.distr_qtys()}}. Can be a
string like \code{'jitter'} (default) or a ggplot2 geom function like
\code{ggplot2::geom_violin}.}

\item{plot_ref_geom}{a geom function to use for the reference plan overlay
in district-level comparison plots, passed to the \code{ref_geom} argument of
\code{\link[redist:redist.plot.distr_qtys]{redist::redist.plot.distr_qtys()}}. Should be a function that accepts \code{...}.}

\item{...}{additional arguments (currently ignored)}
}
\value{
a \code{list}
}
\description{
Set options for \code{redistio}
}
\examples{
redistio_options()
}
